fix: align Expert Profile review patterns
