+++ foundation_cms/templates/patterns/blocks/themes/default/expert_profile_projects_section_block.html
@@ -1,4 +1,4 @@
-{% load i18n responsive_image_tags wagtailimages_tags %}
+{% load i18n wagtailimages_tags %}
 
 {% if rendered_items %}
     <section class="expert-profile-section expert-profile-section--projects">
@@ -19,7 +19,7 @@
                                             width="{{ project_image.width }}"
                                             height="{{ project_image.height }}"
                                         >
-                                        {% if topic %}<a class="expert-profile-project-card__topic" href="{{ topic.get_topic_listing_url }}">{{ topic.name }}</a>{% endif %}
+                                        {% if topic %}<a class="btn-topic expert-profile-project-card__topic" href="{{ topic.get_topic_listing_url }}">{{ topic.name }}</a>{% endif %}
                                     </div>
                                 {% endif %}
                                 <div class="expert-profile-project-card__content">
@@ -30,14 +30,9 @@
                             </article>
                         {% endwith %}
                     {% else %}
-                        <article class="expert-profile-project-card{% if item.value.image %} expert-profile-project-card--with-media{% else %} expert-profile-project-card--text-only{% endif %}">
-                            {% if item.value.image %}
-                                <div class="expert-profile-project-card__media">
-                                    {% responsive_image item.value.image ratio="109:74" base_width=360 sizes="(max-width: 639px) 100vw, 33vw" classnames="expert-profile-project-card__image" %}
-                                </div>
-                            {% endif %}
+                        <article class="expert-profile-project-card expert-profile-project-card--text-only">
                             <div class="expert-profile-project-card__content">
-                                <h3 class="expert-profile-project-card__title"><a href="{{ item.value.url }}" target="_blank" rel="noopener noreferrer">{{ item.value.title }}</a></h3>
+                                <h3 class="expert-profile-project-card__title"><a href="{{ item.value.url }}" target="_blank" rel="noopener noreferrer">{{ item.value.title }} <span class="show-for-sr">{% trans "(opens in a new tab)" %}</span></a></h3>
                                 {% if item.value.description %}<p class="expert-profile-project-card__description">{{ item.value.description }}</p>{% endif %}
                                 <a class="expert-profile-project-card__link expert-profile-project-card__link--external" href="{{ item.value.url }}" target="_blank" rel="noopener noreferrer">
                                     {% if item.value.link_label %}{{ item.value.link_label }}{% else %}{% trans "Learn more" %}{% endif %}
