+++ foundation_cms/templates/patterns/pages/profiles/expert_profile_page.html
@@ -81,7 +81,7 @@
                     <div class="grid-container">
                         <div class="grid-x grid-margin-x">
                             <div class="cell small-12 large-8 large-offset-2">
-                                <h2 class="expert-profile-social__heading" id="expert-profile-social-heading">{% trans "Where to find me?" %}</h2>
+                                <h2 class="expert-profile-social__heading" id="expert-profile-social-heading">{% trans "Where to Find Me" %}</h2>
                                 <ul class="expert-profile-social__list">
                                     {% for social_link in social_links %}
                                         <li>
